+++ src/components/sbb-card/sbb-card.stories.js
@@ -1,10 +1,11 @@
 import { h } from 'jsx-dom';
 import readme from './readme.md';
 
-const Template = (args) => <sbb-card {...args}></sbb-card>;
+const Template = (args) => <sbb-card {...args}>SBB card example text.</sbb-card>;
 
 const TemplateWithBadge = (args) => (
   <sbb-card {...args}>
+    SBB card example text.
     <sbb-card-badge slot="badge" appearance="primary" is-discount></sbb-card-badge>
   </sbb-card>
 );
@@ -47,7 +48,7 @@ sbbCardWithSbbBadge.documentation = {
 export default {
   decorators: [
     (Story) => (
-      <div style={'padding: 2rem'}>
+      <div style={'padding: 2rem; background: lightgray;'}>
         <Story />
       </div>
     ),