+++ src/components/sbb-form-field/sbb-form-field.custom.d.ts
@@ -1,0 +1,4 @@
+export interface InterfaceSbbFormFieldAttributes {
+  errorSpace: 'none' | 'reserve';
+  size: 'l' | 'm';
+}