+++ src/components/sbb-teaser/sbb-teaser.stories.js
@@ -1,6 +1,8 @@
 import { h } from 'jsx-dom';
 import readme from './readme.md';
 
+import placeholderImage from './stories/placeholder.png';
+
 /* ************************************************* */
 /* Storybook controls                                */
 /* ************************************************* */
@@ -60,7 +62,7 @@ const defaultArgTypes = {
   title,
   description,
   'accessibility-label': accessibilityLabel,
-  isStacked,
+  'is-stacked': isStacked,
   href,
 };
 
@@ -69,7 +71,7 @@ const defaultArgs = {
   description: 'This is a paragraph',
   'accessibility-label':
     'The text which gets exposed to screen reader users. The text should reflect all the information which gets passed into the components slots and which is visible in the Teaser, either through text or iconography',
-  isStacked: true,
+  'is-stacked': true,
   href: 'https://github.com/lyne-design-system/lyne-components',
 };
 
@@ -85,7 +87,7 @@ const TemplateDefaultTeaser = (args) => {
   const { title, description, ...remainingArgs } = args;
 
   const sbbTeaserImageArgs = {
-    src: 'https://via.placeholder.com/400x300',
+    src: placeholderImage,
     alt: '400x300 image',
   };
 
@@ -145,7 +147,7 @@ TeaserWithLongText.args = {
 teaserList.argTypes = defaultArgTypes;
 teaserList.args = {
   ...defaultArgs,
-  isStacked: false,
+  'is-stacked': false,
 };
 
 teaserListIsStacked.argTypes = defaultArgTypes;